拍子の分数表示
